% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Highlight.R
\name{Highlightext}
\alias{Highlightext}
\title{Search and highlight text in a PDF
A new copy of the PDF with highlighted text will be created}
\usage{
Highlightext(filepath, filename, text)
}
\arguments{
\item{filepath}{The path where the pdf is located}

\item{filename}{The name of the PDF which is to be searched}

\item{text}{The texts which is to be serached}
}
\description{
Search and highlight text in a PDF
A new copy of the PDF with highlighted text will be created
}
\examples{
Highlightext("/path/to/pdf", "pdf.pdf", c("text1", "text2"))
}
