\name{pdfUtils-package}
\alias{pdfUtils-package}
\alias{pdfUtils}
\docType{package}
\title{
\packageTitle{pdfUtils}
}
\description{
\packageDescription{pdfUtils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pdfUtils}
\packageIndices{pdfUtils}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pdfUtils}

Maintainer: \packageMaintainer{pdfUtils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
